WIP Created new components for the add card button and form
